Rename choropleth2 to choropleth
