+++ src/components-styled/choropleth-tile.tsx
@@ -1,7 +1,7 @@
 import {
   ChoroplethLegenda,
   ILegendaItem,
-} from '~/components/choropleth2/legenda/choropleth-legenda';
+} from '~/components/choropleth/legenda/choropleth-legenda';
 import { useBreakpoints } from '~/utils/useBreakpoints';
 import { Box } from './base';
 import {
