+++ src/components/choropleth/choropleth.module.scss
