+++ src/components/choropleth/hooks/use-chart-dimensions.tsx
