+++ src/components/choropleth/hooks/use-choropleth-color-scale.ts
