+++ src/components/choropleth/hooks/use-municipality-boundingbox.ts
