+++ src/components/choropleth/hooks/use-safety-regionboundingbox.ts
