+++ src/components/choropleth/legenda/choropleth-legenda.module.scss
