+++ src/components/choropleth/legenda/choropleth-legenda.tsx
