+++ src/components/choropleth/legenda/hooks/use-municipal-legenda-data.ts
@@ -1,5 +1,5 @@
-import { municipalThresholds } from '~/components/choropleth2/municipal-thresholds';
-import { TMunicipalityMetricName } from '~/components/choropleth2/shared';
+import { municipalThresholds } from '~/components/choropleth/municipal-thresholds';
+import { TMunicipalityMetricName } from '~/components/choropleth/shared';
 import { useLegendaItems } from './use-legenda-items';
 
 export function useMunicipalLegendaData(metric: TMunicipalityMetricName) {
