+++ src/components/choropleth/legenda/hooks/use-safety-region-legenda-data.ts
