+++ src/components/choropleth/municipal-thresholds.ts
