+++ src/components/choropleth/region-thresholds.ts
