+++ src/components/choropleth/shared.d.ts
