+++ src/components/choropleth/tooltips/municipal/create-municipal-hospital-admissions-tooltip.tsx
@@ -1,8 +1,8 @@
 import { NextRouter } from 'next/router';
 import { ReactNode } from 'react';
-import { MunicipalityProperties } from '~/components/choropleth2/shared';
-import { createSelectMunicipalHandler } from '~/components/choropleth2/select-handlers/create-select-municipal-handler';
-import { TooltipContent } from '~/components/choropleth2/tooltips/tooltipContent';
+import { MunicipalityProperties } from '~/components/choropleth/shared';
+import { createSelectMunicipalHandler } from '~/components/choropleth/select-handlers/create-select-municipal-handler';
+import { TooltipContent } from '~/components/choropleth/tooltips/tooltipContent';
 
 export const createMunicipalHospitalAdmissionsTooltip = (
   router: NextRouter
