+++ src/components/choropleth/tooltips/municipal/create-positive-tested-people-municipal-tooltip.tsx
@@ -2,7 +2,7 @@ import { NextRouter } from 'next/router';
 import { ReactNode } from 'react';
 import { MunicipalityProperties } from '../../shared';
 import { createSelectMunicipalHandler } from '../../select-handlers/create-select-municipal-handler';
-import { TooltipContent } from '~/components/choropleth2/tooltips/tooltipContent';
+import { TooltipContent } from '~/components/choropleth/tooltips/tooltipContent';
 import { replaceVariablesInText } from '~/utils/replaceVariablesInText';
 import siteText from '~/locale/index';
 const text = siteText.common.tooltip;
