+++ src/components/choropleth/tooltips/region/create-infected-locations-regional-tooltip.tsx
@@ -1,6 +1,6 @@
 import { NextRouter } from 'next/router';
 import { ReactNode } from 'react';
-import { TooltipContent } from '~/components/choropleth2/tooltips/tooltipContent';
+import { TooltipContent } from '~/components/choropleth/tooltips/tooltipContent';
 import { RegionsNursingHome } from '~/types/data';
 import { formatNumber, formatPercentage } from '~/utils/formatNumber';
 import { createSelectRegionHandler } from '../../select-handlers/create-select-region-handler';
