+++ src/components/choropleth/tooltips/region/create-region-hospital-admissions-tooltip.tsx
@@ -1,8 +1,8 @@
 import { NextRouter } from 'next/router';
 import { ReactNode } from 'react';
-import { createSelectRegionHandler } from '~/components/choropleth2/select-handlers/create-select-region-handler';
-import { SafetyRegionProperties } from '~/components/choropleth2/shared';
-import { TooltipContent } from '~/components/choropleth2/tooltips/tooltipContent';
+import { createSelectRegionHandler } from '~/components/choropleth/select-handlers/create-select-region-handler';
+import { SafetyRegionProperties } from '~/components/choropleth/shared';
+import { TooltipContent } from '~/components/choropleth/tooltips/tooltipContent';
 import { formatNumber } from '~/utils/formatNumber';
 
 export const createRegionHospitalAdmissionsTooltip = (router: NextRouter) => (
