+++ src/components/choropleth/tooltips/tooltipContent.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import styles from '~/components/choropleth2/tooltips/tooltip.module.scss';
+import styles from '~/components/choropleth/tooltips/tooltip.module.scss';
 
 interface IProps {
   title: string;
