+++ src/pages/landelijk/rioolwater.tsx
@@ -4,10 +4,10 @@ import { Spacer } from '~/components-styled/base';
 import { ChoroplethTile } from '~/components-styled/choropleth-tile';
 import { LineChartTile } from '~/components-styled/line-chart-tile';
 import { ValueAnnotation } from '~/components-styled/value-annotation';
-import { useSafetyRegionLegendaData } from '~/components/choropleth2/legenda/hooks/use-safety-region-legenda-data';
-import { SafetyRegionChoropleth } from '~/components/choropleth2/safety-region-choropleth';
-import { createSelectRegionHandler } from '~/components/choropleth2/select-handlers/create-select-region-handler';
-import { createSewerRegionalTooltip } from '~/components/choropleth2/tooltips/region/create-sewer-regional-tooltip';
+import { useSafetyRegionLegendaData } from '~/components/choropleth/legenda/hooks/use-safety-region-legenda-data';
+import { SafetyRegionChoropleth } from '~/components/choropleth/safety-region-choropleth';
+import { createSelectRegionHandler } from '~/components/choropleth/select-handlers/create-select-region-handler';
+import { createSewerRegionalTooltip } from '~/components/choropleth/tooltips/region/create-sewer-regional-tooltip';
 import { ContentHeader_weekRangeHack } from '~/components/contentHeader_weekRangeHack';
 import { FCWithLayout } from '~/components/layout';
 import { getNationalLayout } from '~/components/layout/NationalLayout';
