+++ src/pages/veiligheidsregio/[code]/ziekenhuis-opnames.tsx
@@ -1,11 +1,11 @@
 import { useRouter } from 'next/router';
 import Ziekenhuis from '~/assets/ziekenhuis.svg';
 import { LineChart } from '~/components/charts/index';
-import { ChoroplethLegenda } from '~/components/choropleth2/legenda/choropleth-legenda';
-import { useMunicipalLegendaData } from '~/components/choropleth2/legenda/hooks/use-municipal-legenda-data';
-import { MunicipalityChoropleth } from '~/components/choropleth2/municipality-choropleth';
-import { createSelectMunicipalHandler } from '~/components/choropleth2/select-handlers/create-select-municipal-handler';
-import { createMunicipalHospitalAdmissionsTooltip } from '~/components/choropleth2/tooltips/municipal/create-municipal-hospital-admissions-tooltip';
+import { ChoroplethLegenda } from '~/components/choropleth/legenda/choropleth-legenda';
+import { useMunicipalLegendaData } from '~/components/choropleth/legenda/hooks/use-municipal-legenda-data';
+import { MunicipalityChoropleth } from '~/components/choropleth/municipality-choropleth';
+import { createSelectMunicipalHandler } from '~/components/choropleth/select-handlers/create-select-municipal-handler';
+import { createMunicipalHospitalAdmissionsTooltip } from '~/components/choropleth/tooltips/municipal/create-municipal-hospital-admissions-tooltip';
 import { DataWarning } from '~/components/dataWarning';
 import { FCWithLayout } from '~/components/layout';
 import { ContentHeader } from '~/components/contentHeader';
