+++ src/pages/veiligheidsregio/index.tsx
@@ -5,12 +5,12 @@ import EscalationLevel1 from '~/assets/niveau-1.svg';
 import EscalationLevel2 from '~/assets/niveau-2.svg';
 import EscalationLevel3 from '~/assets/niveau-3.svg';
 import EscalationLevel4 from '~/assets/niveau-4.svg';
-import { regionThresholds } from '~/components/choropleth2/region-thresholds';
-import { SafetyRegionChoropleth } from '~/components/choropleth2/safety-region-choropleth';
-import { createSelectRegionHandler } from '~/components/choropleth2/select-handlers/create-select-region-handler';
-import { ChoroplethThresholds } from '~/components/choropleth2/shared';
-import { escalationTooltip } from '~/components/choropleth2/tooltips/region/escalation-tooltip';
-import styles from '~/components/choropleth2/tooltips/tooltip.module.scss';
+import { regionThresholds } from '~/components/choropleth/region-thresholds';
+import { SafetyRegionChoropleth } from '~/components/choropleth/safety-region-choropleth';
+import { createSelectRegionHandler } from '~/components/choropleth/select-handlers/create-select-region-handler';
+import { ChoroplethThresholds } from '~/components/choropleth/shared';
+import { escalationTooltip } from '~/components/choropleth/tooltips/region/escalation-tooltip';
+import styles from '~/components/choropleth/tooltips/tooltip.module.scss';
 import { FCWithLayout } from '~/components/layout';
 import { getSafetyRegionLayout } from '~/components/layout/SafetyRegionLayout';
 import { SEOHead } from '~/components/seoHead';
